requirements.txt
pandas
requests
python-dotenv
tenacity
openpyxl
